+++ src/contexts/ThemeContext.tsx
@@ -1,0 +1,40 @@
+import {
+  createContext,
+  ReactNode,
+  useCallback,
+  useContext,
+  useState,
+} from "react";
+
+export interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(
+  undefined,
+);
+
+type Props = {
+  children: ReactNode;
+};
+
+export const ThemeProvider = ({ children }: Props) => {
+  const [theme, setTheme] = useState("light");
+  const toggleTheme = useCallback(() => {
+    setTheme((prevTheme) => (prevTheme === "light" ? "dark" : "light"));
+  }, []);
+  return (
+    <ThemeContext.Provider value={{ theme, toggleTheme }}>
+      {children}
+    </ThemeContext.Provider>
+  );
+};
+
+export const useThemeContext = () => {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error("useThemeContext must be used within an AppProvider");
+  }
+  return context;
+};
