+++ src/contexts/UserContext.tsx
@@ -1,0 +1,53 @@
+import { createContext, ReactNode, useContext, useState } from "react";
+import { User } from "../types/user";
+import { useCallback, useMemo } from "../@lib";
+import { useNotificationContext } from "./NotificationContext";
+
+export interface UserContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
+
+export const UserContext = createContext<UserContextType | undefined>(
+  undefined,
+);
+
+type Props = {
+  children: ReactNode;
+};
+
+export const UserProvider = ({ children }: Props) => {
+  const [user, setUser] = useState<User | null>(null);
+  const { addNotification } = useNotificationContext();
+
+  const login = useCallback(
+    (email: string) => {
+      setUser({ id: 1, name: "홍길동", email });
+      addNotification("성공적으로 로그인되었습니다", "success");
+    },
+    [addNotification],
+  );
+
+  const logout = useCallback(() => {
+    setUser(null);
+    addNotification("로그아웃되었습니다", "info");
+  }, [addNotification]);
+
+  const contextValue = useMemo(
+    () => ({ user, login, logout }),
+    [user, login, logout],
+  );
+
+  return (
+    <UserContext.Provider value={contextValue}>{children}</UserContext.Provider>
+  );
+};
+
+export const useUserContext = () => {
+  const context = useContext(UserContext);
+  if (context === undefined) {
+    throw new Error("useUserContext must be used within an AppProvider");
+  }
+  return context;
+};
