+++ src/contexts/app.ts
@@ -1,0 +1,16 @@
+import { createContext, useContext } from "react";
+
+export interface AppContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
+
+export const AppContext = createContext<AppContextType | undefined>(undefined);
+
+export const useAppContext = () => {
+  const context = useContext(AppContext);
+  if (context === undefined) {
+    throw new Error("useAppContext must be used within an AppProvider");
+  }
+  return context;
+};
